Use the Meson build system.
These days, using Meson is considered good practice for projects
using GNOME technology. It is well integrated into GNOME Builder.
Compared to Makefiles, Meson build files are more portable and
packagers can expect a common interface.

All the source files are moved into a subdirectory. This is
done purely for aestethic reasons and may as well be moved back
into the root directory.

.gitignore is removed because the build is typically not done
in the source directory. The PKGBUILD is removed, because it is
specific to Arch Linux and should better be done in their AUR.

The meson.build files specify build targets and tell what to
install:

* /meson.build only includes other build files.
* /src/meson.build creates and installs the shared library,
  along with the C header, GIR and Typelib.
* /test/meson.build creates a test binary, which can be
  executed in the build directory using `ninja test`.